% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_auth.R
\name{app_username}
\alias{app_username}
\title{Whitelist your app and receive a username}
\usage{
app_username(token, app_id = Sys.getenv("PHILIPS_HUE_APP_ID"))
}
\arguments{
\item{token}{token returned from \code{\link[=token]{token()}}}

\item{app_id}{your app's ID (the name you registered your app under)}
}
\value{
Adds your app ID to the list of whitelisted apps on the user's bridge
and returns a username that can be used to authenticate future requests.
}
\description{
Whitelist your app and receive a username
}
