% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_local}
\alias{auth_local}
\title{Authentication -- local}
\usage{
auth_local(ip = Sys.getenv("PHILIPS_HUE_BRIDGE_IP"),
  username = Sys.getenv("PHILIPS_HUE_BRIDGE_USERNAME"))
}
\arguments{
\item{ip}{the IP address of your Hue Bridge}

\item{username}{the username with access to your Hue Bridge}
}
\value{
Returns \code{TRUE} (invisibly) if options were successfully set
}
\description{
This function helps check and set the necessary environment variables to
authenticate to a Hue Bridge on the local network.
}
\seealso{
\url{https://www.developers.meethue.com/documentation/getting-started}
}
