% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_remote}
\alias{auth_remote}
\title{Authentication -- remote}
\usage{
auth_remote(app_id = Sys.getenv("PHILIPS_HUE_APP_ID"),
  client_id = Sys.getenv("PHILIPS_HUE_CLIENT_ID"),
  client_secret = Sys.getenv("PHILIPS_HUE_CLIENT_SECRET"),
  bridge_id = Sys.getenv("PHILIPS_HUE_BRIDGE_ID"),
  bridge_name = Sys.getenv("PHILIPS_HUE_BRIDGE_NAME"),
  username = Sys.getenv("PHILIPS_HUE_BRIDGE_REMOTE_USERNAME"),
  access_token = Sys.getenv("PHILIPS_HUE_ACCESS_TOKEN"),
  access_token_exp = Sys.getenv("PHILIPS_HUE_ACCESS_TOKEN_EXP"),
  refresh_token = Sys.getenv("PHILIPS_HUE_REFRESH_TOKEN"),
  refresh_token_exp = Sys.getenv("PHILIPS_HUE_REFRESH_TOKEN_EXP"),
  initial_setup = interactive())
}
\arguments{
\item{app_id}{your app's ID (the name you registered your app under)}

\item{client_id}{your app's client ID (assigned when you registered your app
with Hue)}

\item{client_secret}{your app's client secret (assigned when you registered
your app with Hue)}

\item{bridge_id}{ID of hue bridge you'd like remote access to}

\item{bridge_name}{name of hue bridge you'd like remote access to}

\item{username}{TODO}

\item{access_token}{access token returned from \code{\link[=request_token]{request_token()}} (or a
subsequent \code{\link[=refresh_token]{refresh_token()}})}

\item{access_token_exp}{access token expiration time returned from
\code{\link[=request_token]{request_token()}} (or a subsequent \code{\link[=refresh_token]{refresh_token()}})}

\item{refresh_token}{refresh token returned from \code{\link[=request_token]{request_token()}} (or a
subsequent \code{\link[=refresh_token]{refresh_token()}})}

\item{refresh_token_exp}{refresh token expiration time returned from
\code{\link[=request_token]{request_token()}} (or a subsequent \code{\link[=refresh_token]{refresh_token()}})}

\item{initial_setup}{if \code{TRUE}, and initial app authorization is required,
\code{\link[=auth_remote]{auth_remote()}} will prompt the user to authorize the app and enter the
auth code manually}
}
\value{
Returns \code{TRUE} (invisibly) if options were successfully set
}
\description{
TODO
}
\seealso{
\url{https://developers.meethue.com/develop/hue-api/remote-api-quick-start-guide/}
}
