% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bridge_url}
\alias{bridge_url}
\title{Create a Hue API endpoint URL}
\usage{
bridge_url(...)
}
\arguments{
\item{...}{strings to append to the root URL}
}
\value{
Returns a complete URL for the API endpoint.
}
\description{
This function creates a URL for a Hue API endpoint. The base URL is created
using secrets set with a call to \code{\link[=set_bridge_credentials]{set_bridge_credentials()}}. It
then appends any user-supplied components, separating each with a "/".
}
\seealso{
\code{\link[=set_bridge_credentials]{set_bridge_credentials()}}
}
