% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities_api.R
\name{capabilities}
\alias{capabilities}
\alias{get_capabilities}
\title{Hue API: \code{capabilities} endpoints}
\usage{
get_capabilities()
}
\value{
\link{get_capabilities} returns a list structure with the capabilities of
bridge resources.
}
\description{
Hue API: \code{capabilities} endpoints
}
\seealso{
\url{https://developers.meethue.com/develop/hue-api/10-capabilities-api/}
}
