% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_api.R
\name{configs}
\alias{configs}
\alias{get_config}
\alias{set_config_attributes}
\alias{get_state}
\title{Hue API: \code{configuration} endpoints}
\usage{
get_config()

set_config_attributes(...)

get_state()
}
\arguments{
\item{...}{named parameters describing configuration}
}
\value{
Requests that create resources return the ID of the newly created
item, requests with side effects return \code{TRUE} upon success, and GET
requests return the response content, parsed into a list.
}
\description{
Hue API: \code{configuration} endpoints
}
\seealso{
\url{https://developers.meethue.com/develop/hue-api/7-configuration-api/}
}
