% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{configure_daylight_sensor}
\alias{configure_daylight_sensor}
\title{Configure Built-In Daylight Sensor}
\usage{
configure_daylight_sensor(lat, lon, sunriseoffset = 30, sunsetoffset = -30,
  id = 1)
}
\arguments{
\item{lat}{latitude}

\item{lon}{logitude}

\item{sunriseoffset}{"daylight" begins \code{sunriseoffset} minutes after
sunrise}

\item{sunsetoffset}{"daylight" ends \code{sunsetoffset} minutes after sunset}

\item{id}{ID of the daylight sensor}
}
\value{
Returns \code{TRUE} (invisibly) uppon success
}
\description{
Configure Built-In Daylight Sensor
}
