% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_api.R
\name{create_user}
\alias{create_user}
\title{Create local user}
\usage{
create_user(devicetype)
}
\arguments{
\item{devicetype}{a string naming your app and the device it's running on;
suggested format: \verb{<application_name>#<devicename>} (e.g. \code{Hue#iPhone13}).}
}
\value{
Returns a list with the newly created \code{username} and \code{clientkey}
}
\description{
The \code{\link[=create_user]{create_user()}} function allows you to create a user on a local Hue
network, but it requires pressing the button on the Hue bridge then
executing this command within 30 seconds. \code{delete_user()} is not implemented
because, apparently, it can only be executed through remote authentication.
}
\seealso{
\url{https://developers.meethue.com/develop/hue-api/7-configuration-api/#create-user}
}
