% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capabilities.R
\name{get_capabilities}
\alias{get_capabilities}
\title{Get Capabilities}
\usage{
get_capabilities()
}
\value{
Returns a list structure with the capabilities of bridges resources.
}
\description{
Get Capabilities
}
