% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups}
\alias{groups}
\alias{create_group}
\alias{get_groups}
\alias{get_group}
\alias{set_group_attributes}
\alias{set_group_state}
\alias{delete_group}
\title{Hue API: groups endpoints}
\usage{
create_group(..., return_id = FALSE)

get_groups()

get_group(id)

set_group_attributes(id, ...)

set_group_state(id, ...)

delete_group(id)
}
\arguments{
\item{...}{named parameters describing group attributes or state
(e.g. \code{name = 'foo'}; \code{on = TRUE})}

\item{return_id}{when creating a new resource, the default is to return
\code{TRUE} (invisibly) upon success; setting \code{return_id = TRUE} will
return the ID of the newly created resource instead.}

\item{id}{id of a specific group}
}
\value{
Requests with side effects return \code{TRUE} (invisibly) upon
success. GET requests return the response content, parsed into a list.
}
\description{
Hue API: groups endpoints
}
\seealso{
\url{https://developers.meethue.com/documentation/groups-api}
}
