% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{guess_room_class}
\alias{guess_room_class}
\title{Guess Room Class}
\usage{
guess_room_class(x)
}
\arguments{
\item{x}{room name}
}
\value{
Returns a character vector with the best guess at the room class of
  the given room name
}
\description{
Guess Room Class
}
