% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lights.R
\name{lights}
\alias{lights}
\alias{search_for_new_lights}
\alias{get_new_lights}
\alias{rename_light}
\alias{get_lights}
\alias{get_light}
\alias{set_light_state}
\alias{delete_light}
\title{Hue API: light endpoints}
\usage{
search_for_new_lights()

get_new_lights()

rename_light(id, name)

get_lights()

get_light(id)

set_light_state(id, ...)

delete_light(id)
}
\arguments{
\item{id}{id of a specific light, an integer}

\item{name}{name to give to a specific light}

\item{...}{named parameters describing light state}
}
\value{
Returns the response content, conveniently parsed into an R-friendly
  data format.
}
\description{
Hue API: light endpoints
}
