% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhilipsHue-package.R
\docType{package}
\name{PhilipsHue-package}
\alias{PhilipsHue}
\alias{PhilipsHue-package}
\title{PhilipsHue: R Interface to the Philips Hue API}
\description{
The PhilipsHue package provides an R interface to the Philips Hue
API service, making it easy to make API calls and process the response into
convenient, R-friendly data structures. See
<https://developers.meethue.com/philips-hue-api> for the Philips Hue API
Documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/fascinatingfingers/philipshue}
  \item Report bugs at \url{https://gitlab.com/fascinatingfingers/philipshue/issues}
}

}
\author{
\strong{Maintainer}: Justin Brantley \email{fascinatingfingers@icloud.com}

}
\keyword{internal}
