% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{process_httr_response}
\alias{process_httr_response}
\title{Process HTTP Responses}
\usage{
process_httr_response(x, ...)
}
\arguments{
\item{x}{an \code{httr} \code{\link[httr]{response}} object}

\item{...}{parameters passed to \code{\link[jsonlite]{fromJSON}}}
}
\value{
Returns the content of the response, parsed into an R-friendly list.
}
\description{
This function performs basic error handling and parsing of \code{httr}
  \code{\link[httr]{response}} objects.
}
