% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{refresh_token}
\alias{refresh_token}
\title{Refresh access token}
\usage{
refresh_token(refresh_token = Sys.getenv("PHILIPS_HUE_REFRESH_TOKEN"),
  client_id = Sys.getenv("PHILIPS_HUE_CLIENT_ID"),
  client_secret = Sys.getenv("PHILIPS_HUE_CLIENT_SECRET"))
}
\arguments{
\item{refresh_token}{refresh token returned from \code{\link[=request_token]{request_token()}} (or
a subsequent \code{\link[=refresh_token]{refresh_token()}})}

\item{client_id}{your app's client ID (assigned when you registered your app
with Hue)}

\item{client_secret}{your app's client secret (assigned when you registered
your app with Hue)}
}
\value{
If successful, returns a list with authorization/refresh tokens and
expiration times
}
\description{
Tokens received by calling \code{\link[=request_token]{request_token()}} expire but can be refreshed by
a simple call to \code{\link[=refresh_token]{refresh_token()}}.
}
