% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{remote_auth}
\alias{remote_auth}
\title{Remotely activate a button press on the user's bridge}
\usage{
remote_auth(token)
}
\arguments{
\item{token}{token returned from \code{\link[=request_token]{request_token()}}}
}
\value{
Returns \code{TRUE} (invisibly) if the call was successful.
}
\description{
After obtaining an authorization token (see \code{\link[=request_token]{request_token()}}), the token can
be used to remotely "press" the button on the Hue bridge as part of a remote
authentication sequence.
}
