% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_auth.R
\name{remote_auth_request_url}
\alias{remote_auth_request_url}
\title{Request remote access to user's bridge}
\usage{
remote_auth_request_url(client_id = Sys.getenv("PHILIPS_HUE_CLIENT_ID"),
  app_id = Sys.getenv("PHILIPS_HUE_APP_ID"),
  bridge_id = Sys.getenv("PHILIPS_HUE_BRIDGE_ID"),
  bridge_name = Sys.getenv("PHILIPS_HUE_BRIDGE_NAME"), state = NULL)
}
\arguments{
\item{client_id}{your app's client ID (assigned when you registered your app
with Hue)}

\item{app_id}{your app's ID (the name you registered your app under)}

\item{bridge_id}{ID of hue bridge you'd like remote access to}

\item{bridge_name}{name of hue bridge you'd like remote access to}

\item{state}{arbitrary data that will be included in the redirected response
(see Details)

When \code{state = NULL} (the default), a UUID will be generated. You can verify
this value in the redirected response for an added layer of security.}
}
\value{
Returns a URL to request remote access to a user's bridge.
}
\description{
The first step in the remote auth process is to request remote access to the
user's bridge. When they click the generated URL, they will be taken to the
Hue website where they can grant remote access to your app. Once approved,
Hue will redirect them to the callback URL you provided when you registered
your app. This redirect URL will include an authorization code as a query
parameter. Retain this code.
}
\examples{
remote_auth_request_url('client_id', 'app id', 'bridge_id', 'bridge name')

}
