% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_auth.R
\name{remote_button_press}
\alias{remote_button_press}
\title{Remotely activate a button press on the user's bridge}
\usage{
remote_button_press(token)
}
\arguments{
\item{token}{token returned from \code{\link[=token]{token()}}}
}
\value{
Returns \code{TRUE} (invisibly) if the call was successful.
}
\description{
After obtaining an authorization token (see \code{\link[=token]{token()}}), the token can be
used to remotely "press" the button on the Hue bridge as part of a remote
authentication sequence.
}
