% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_auth.R
\name{request_token}
\alias{request_token}
\title{Request authorization token}
\usage{
request_token(auth_code = Sys.getenv("PHILIPS_HUE_AUTH_CODE"),
  client_id = Sys.getenv("PHILIPS_HUE_CLIENT_ID"),
  client_secret = Sys.getenv("PHILIPS_HUE_CLIENT_SECRET"))
}
\arguments{
\item{auth_code}{authorization code sent to your app's callback URL (see
\code{\link[=authorize_at]{authorize_at()}})}

\item{client_id}{your app's client ID (assigned when you registered your app
with Hue)}

\item{client_secret}{your app's client secret (assigned when you registered
your app with Hue)}
}
\value{
If successful, returns a list with authorization/refresh tokens and
expiration times
}
\description{
After you obtain an authorization code (see \code{\link[=authorize_at]{authorize_at()}}), use the code
to obtain an authorization token.
}
