% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{reset_auth}
\alias{reset_auth}
\title{Reset authentication environment variables}
\usage{
reset_auth(local = TRUE, remote = TRUE)
}
\arguments{
\item{local}{if \code{TRUE}, resets environment variables required for local
authentication}

\item{remote}{if \code{TRUE}, resets environment variables required for remote
authentication}
}
\value{
Returns \code{TRUE} invisibly upon success
}
\description{
Reset authentication environment variables
}
