% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcelinks.R
\name{resourcelinks}
\alias{resourcelinks}
\alias{create_resourcelink}
\alias{get_resourcelinks}
\alias{get_resourcelink}
\alias{set_resourcelink_attributes}
\alias{delete_resourcelink}
\title{Hue API: resourcelinks endpoints}
\usage{
create_resourcelink(...)

get_resourcelinks()

get_resourcelink(id)

set_resourcelink_attributes(id, ...)

delete_resourcelink(id)
}
\arguments{
\item{...}{named parameters describing resourcelink attributes
(e.g. \code{name = 'foo'})}

\item{id}{id of a specific resourcelink, an integer}
}
\value{
Requests with side effects invisibly return \code{TRUE} upon success.
  GET requests return the response content, parsed into a list.
}
\description{
Hue API: resourcelinks endpoints
}
\seealso{
\url{https://developers.meethue.com/documentation/resourcelinks-api}
}
