% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rule_helpers}
\alias{rule_helpers}
\alias{condition}
\alias{action}
\title{Rule Helpers}
\usage{
condition(address, operator, value = NULL)

action(address, method, ...)
}
\arguments{
\item{address}{path to attribute or resource}

\item{operator}{one of: eq, gt, lt, dx, ddx, stable, not stable, in, not in}

\item{value}{the value a condition will compare against}

\item{method}{the HTTP method used to send the body to the given address}

\item{...}{named parameters to include in action body}
}
\value{
Returns a list-like structure suitable for \code{\link{create_rule}}
  or \code{\link{set_rule_attributes}}.
}
\description{
Defining rules can become quite verbose, and it can be tricky to prepare the
proper list structure for the POST or PUT request. These functions simplify
things a bit and provide a leaner, more sematic interface.
}
