% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rules}
\alias{rules}
\alias{create_rule}
\alias{get_rules}
\alias{get_rule}
\alias{set_rule_attributes}
\alias{delete_rule}
\title{Hue API: rules endpoints}
\usage{
create_rule(name, conditions, actions)

get_rules()

get_rule(id)

set_rule_attributes(id, name = NULL, conditions = NULL, actions = NULL)

delete_rule(id)
}
\arguments{
\item{name}{a name to give the rule}

\item{conditions}{a list of conditions (e.g. the result of a call to
\code{\link{condition}} )}

\item{actions}{a list of conditions (e.g. the result of a call to
\code{\link{action}} )}

\item{id}{id of a specific rule, an integer}
}
\value{
Requests with side effects invisibly return \code{TRUE} upon success.
  GET requests return the response content, parsed into a list.
}
\description{
Hue API: rules endpoints
}
\seealso{
\url{https://developers.meethue.com/documentation/rules-api}
}
