% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedules.R
\name{schedules}
\alias{schedules}
\alias{create_schedule}
\alias{get_schedules}
\alias{get_schedule}
\alias{set_schedule_attributes}
\alias{delete_schedule}
\title{Hue API: schedules endpoints}
\usage{
create_schedule(...)

get_schedules()

get_schedule(id)

set_schedule_attributes(id, ...)

delete_schedule(id)
}
\arguments{
\item{...}{named parameters describing schedule attributes
(e.g. \code{name = 'foo'})}

\item{id}{id of a specific schedule, an integer}
}
\value{
Requests with side effects invisibly return \code{TRUE} upon success.
  GET requests return the response content, parsed into a list.
}
\description{
Hue API: schedules endpoints
}
\seealso{
\url{https://developers.meethue.com/documentation/schedules-api-0}
}
