% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_helpers.R
\name{set_bridge_credentials}
\alias{set_bridge_credentials}
\alias{reset_bridge_credentials}
\title{Hue Bridge Authentication}
\usage{
set_bridge_credentials(ip, username)

reset_bridge_credentials()
}
\arguments{
\item{ip}{the IP address of your Hue Bridge}

\item{username}{the username with access to your Hue Bridge}
}
\value{
Returns \code{TRUE} (invisibly) if a connection to the Hue Bridge was
  successful.
}
\description{
These functions are used primarily for their side effects -- specifically to
set (or reset) Hue Bridge authentication secrets stored in
\code{options()$PhilipsHue}.
}
\seealso{
\url{https://www.developers.meethue.com/documentation/getting-started}
}
