% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_auth.R
\name{token}
\alias{token}
\title{Request authorization token}
\usage{
token(auth_code = Sys.getenv("PHILIPS_HUE_AUTH_CODE"),
  client_id = Sys.getenv("PHILIPS_HUE_CLIENT_ID"),
  client_secret = Sys.getenv("PHILIPS_HUE_CLIENT_SECRET"))
}
\arguments{
\item{auth_code}{authorization code sent to your app's callback URL
(see \code{\link[=remote_auth_request_url]{remote_auth_request_url()}})}

\item{client_id}{your app's client ID (assigned when you registered your app
with Hue)}

\item{client_secret}{your app's client secret (assigned when you registered
your app with Hue)}
}
\value{
If successful, returns a list with authorization/refresh tokens and
expiration times
}
\description{
After you obtain an authorization code (see \code{\link[=remote_auth_request_url]{remote_auth_request_url()}}), use
the code to obtain an authorization token.
}
