% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_helpers.R
\name{bridge_url}
\alias{bridge_url}
\title{Create a Hue API endpoint URL}
\usage{
bridge_url(...)
}
\arguments{
\item{...}{strings to append to the root URL}
}
\value{
Returns a complete URL for the API endpoint
}
\description{
Create a Hue API endpoint URL
}
\seealso{
\code{\link{set_bridge_credentials}}
}
