% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_api.R
\name{groups}
\alias{groups}
\alias{create_group}
\alias{get_groups}
\alias{get_group}
\alias{set_group_attributes}
\alias{set_group_state}
\alias{delete_group}
\title{Hue API: \code{groups} endpoints}
\usage{
create_group(...)

get_groups()

get_group(id)

set_group_attributes(id, ...)

set_group_state(id, ...)

delete_group(id)
}
\arguments{
\item{...}{named parameters describing group attributes or state
(e.g. \code{name = "foo"}; \code{on = TRUE})}

\item{id}{ID of a specific group}
}
\value{
Requests that create resources return the ID of the newly created
item, requests with side effects return \code{TRUE} upon success, and GET
requests return the response content, parsed into a list.
}
\description{
Hue API: \code{groups} endpoints
}
\seealso{
\url{https://developers.meethue.com/develop/hue-api/groupds-api/}
}
