% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{guess_room_class}
\alias{guess_room_class}
\title{Guess Room Class}
\usage{
guess_room_class(x)
}
\arguments{
\item{x}{room name}
}
\value{
Returns a character vector with the single best guess of the room
  class of the given room name.
}
\description{
Every new room must be assigned a room "class" (Living room, Kitchen, etc).
This function attempts to guess the room class from the room name. For
example, a room named "Master Bedroom" would be assigned the class "Bedroom"
because it contains a substring match.
}
