% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lights_api.R
\name{lights}
\alias{lights}
\alias{search_for_new_lights}
\alias{get_new_lights}
\alias{rename_light}
\alias{get_lights}
\alias{get_light}
\alias{set_light_state}
\alias{delete_light}
\title{Hue API: \code{lights} endpoints}
\usage{
search_for_new_lights()

get_new_lights()

rename_light(id, name)

get_lights()

get_light(id)

set_light_state(id, ...)

delete_light(id)
}
\arguments{
\item{id}{ID of a specific light}

\item{name}{name to assign to the light}

\item{...}{named parameters describing light state (e.g. \code{on = TRUE})}
}
\value{
Requests that create resources return the ID of the newly created
item, requests with side effects return \code{TRUE} upon success, and GET
requests return the response content, parsed into a list.
}
\description{
Hue API: \code{lights} endpoints
}
\seealso{
\url{https://developers.meethue.com/develop/hue-api/lights-api/}
}
