% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lights.R
\name{lights}
\alias{lights}
\alias{search_for_new_lights}
\alias{get_new_lights}
\alias{rename_light}
\alias{get_lights}
\alias{get_light}
\alias{set_light_state}
\alias{delete_light}
\title{Hue API: lights endpoints}
\usage{
search_for_new_lights()

get_new_lights()

rename_light(id, name)

get_lights()

get_light(id)

set_light_state(id, ...)

delete_light(id)
}
\arguments{
\item{id}{id of a specific light, an integer}

\item{name}{name to give to a specific light}

\item{...}{named parameters describing light state (e.g. \code{on = TRUE})}
}
\value{
Requests with side effects invisibly return \code{TRUE} upon success.
  GET requests return the response content, parsed into a list.
}
\description{
Hue API: lights endpoints
}
\seealso{
\url{https://developers.meethue.com/documentation/lights-api}
}
