% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhilipsHue-package.R
\docType{package}
\name{PhilipsHue-package}
\alias{PhilipsHue}
\alias{PhilipsHue-package}
\title{PhilipsHue: R Interface to the Philips Hue API}
\description{
Control Philips Hue smart lighting. Use this package to connect to a Hue bridge on your local network (remote authentication not yet supported) and control your smart lights through the Philips Hue API. All API V1 endpoints are supported. See API documentation at <https://developers.meethue.com/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fascinatingfingers.gitlab.io/philipshue}
  \item \url{https://gitlab.com/fascinatingfingers/philipshue}
  \item Report bugs at \url{https://gitlab.com/fascinatingfingers/philipshue/-/issues}
}

}
\author{
\strong{Maintainer}: Justin Brantley \email{fascinatingfingers@icloud.com}

}
\keyword{internal}
