% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_links_api.R
\name{resourcelinks}
\alias{resourcelinks}
\alias{create_resourcelink}
\alias{get_resourcelinks}
\alias{get_resourcelink}
\alias{set_resourcelink_attributes}
\alias{delete_resourcelink}
\title{Hue API: \code{resourcelinks} endpoints}
\usage{
create_resourcelink(...)

get_resourcelinks()

get_resourcelink(id)

set_resourcelink_attributes(id, ...)

delete_resourcelink(id)
}
\arguments{
\item{...}{named parameters describing resourcelink attributes
(e.g. \code{name = "foo"})}

\item{id}{ID of a specific resourcelink}
}
\value{
Requests that create resources return the ID of the newly created
item, requests with side effects return \code{TRUE} upon success, and GET
requests return the response content, parsed into a list.
}
\description{
Hue API: \code{resourcelinks} endpoints
}
\seealso{
\url{https://developers.meethue.com/develop/hue-api/9-resourcelinks-api/}
}
