% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcelinks.R
\name{resourcelinks}
\alias{resourcelinks}
\alias{create_resourcelink}
\alias{get_resourcelinks}
\alias{get_resourcelink}
\alias{set_resourcelink_attributes}
\alias{delete_resourcelink}
\title{Hue API: resourcelinks endpoints}
\usage{
create_resourcelink(..., return_id = FALSE)

get_resourcelinks()

get_resourcelink(id)

set_resourcelink_attributes(id, ...)

delete_resourcelink(id)
}
\arguments{
\item{...}{named parameters describing resourcelink attributes
(e.g. \code{name = 'foo'})}

\item{return_id}{when creating a new resource, the default is to return
\code{TRUE} (invisibly) upon success; setting \code{return_id = TRUE} will
return the ID of the newly created resource instead.}

\item{id}{id of a specific resourcelink}
}
\value{
Requests with side effects return \code{TRUE} (invisibly) upon
success. GET requests return the response content, parsed into a list.
}
\description{
Hue API: resourcelinks endpoints
}
\seealso{
\url{https://developers.meethue.com/documentation/resourcelinks-api}
}
