% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_api.R
\name{rules}
\alias{rules}
\alias{create_rule}
\alias{get_rules}
\alias{get_rule}
\alias{set_rule_attributes}
\alias{delete_rule}
\title{Hue API: \code{rules} endpoints}
\usage{
create_rule(name, conditions, actions)

get_rules()

get_rule(id)

set_rule_attributes(id, name, conditions, actions)

delete_rule(id)
}
\arguments{
\item{name}{name to assign to the rule}

\item{conditions}{a list of conditions (e.g. the result of a call to
\code{\link[=condition]{condition()}} )}

\item{actions}{a list of actions (e.g. the result of a call to
\code{\link[=action]{action()}} )}

\item{id}{ID of a specific rule}
}
\value{
Requests that create resources return the ID of the newly created
item, requests with side effects return \code{TRUE} upon success, and GET
requests return the response content, parsed into a list.
}
\description{
Hue API: \code{rules} endpoints
}
\seealso{
\url{https://developers.meethue.com/develop/hue-api/6-rules-api/}
}
