% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rules}
\alias{rules}
\alias{create_rule}
\alias{get_rules}
\alias{get_rule}
\alias{set_rule_attributes}
\alias{delete_rule}
\title{Hue API: rules endpoints}
\usage{
create_rule(name, conditions, actions, return_id = FALSE)

get_rules()

get_rule(id)

set_rule_attributes(id, name = NULL, conditions = NULL,
  actions = NULL)

delete_rule(id)
}
\arguments{
\item{name}{name to assign to the rule}

\item{conditions}{a list of conditions (e.g. the result of a call to
\code{\link[=condition]{condition()}} )}

\item{actions}{a list of actions (e.g. the result of a call to
\code{\link[=action]{action()}} )}

\item{return_id}{when creating a new resource, the default is to return
\code{TRUE} (invisibly) upon success; setting \code{return_id = TRUE} will
return the ID of the newly created resource instead.}

\item{id}{id of a specific rule}
}
\value{
Requests with side effects return \code{TRUE} (invisibly) upon
success. GET requests return the response content, parsed into a list.
}
\description{
Hue API: rules endpoints
}
\seealso{
\url{https://developers.meethue.com/documentation/rules-api}
}
