% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenes.R
\name{scenes}
\alias{scenes}
\alias{create_scene}
\alias{get_scenes}
\alias{get_scene}
\alias{set_scene_attributes}
\alias{set_scene_lightstate}
\alias{delete_scene}
\title{Hue API: scenes endpoints}
\usage{
create_scene(name, lights, recycle = TRUE, transitiontime = 4,
  return_id = FALSE)

get_scenes()

get_scene(id)

set_scene_attributes(id, ...)

set_scene_lightstate(scene_id, light_id, ...)

delete_scene(id)
}
\arguments{
\item{name}{name for the scene}

\item{lights}{vector of light IDs included in the scene}

\item{recycle}{logical indicating whether the scene can be automatically
deleted by the bridge}

\item{transitiontime}{duration (in milliseconds) of the scene transition}

\item{return_id}{if \code{TRUE} returns the ID of the newly created object
rather than \code{invisible(TRUE)}}

\item{id, scene_id}{id of a specific scene}

\item{...}{named parameters describing scene attributes or lightstate
(e.g. \code{name = 'foo'}; \code{on = TRUE})}

\item{light_id}{id of a specific light in the scene}
}
\value{
Requests with side effects invisibly return \code{TRUE} upon success.
  GET requests return the response content, parsed into a list.
}
\description{
Hue API: scenes endpoints
}
\seealso{
\url{https://developers.meethue.com/documentation/scenes-api}
}
