% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedules_api.R
\name{schedules}
\alias{schedules}
\alias{create_schedule}
\alias{get_schedules}
\alias{get_schedule}
\alias{set_schedule_attributes}
\alias{delete_schedule}
\title{Hue API: \code{schedules} endpoints}
\usage{
create_schedule(...)

get_schedules()

get_schedule(id)

set_schedule_attributes(id, ...)

delete_schedule(id)
}
\arguments{
\item{...}{named parameters describing schedule attributes
(e.g. \code{name = "foo"})}

\item{id}{ID of a specific schedule}
}
\value{
Requests that create resources return the ID of the newly created
item, requests with side effects return \code{TRUE} upon success, and GET
requests return the response content, parsed into a list.
}
\description{
Hue API: \code{schedules} endpoints
}
\seealso{
\url{https://developers.meethue.com/develop/hue-api/3-schedules-api/}
}
