% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedules.R
\name{schedules}
\alias{schedules}
\alias{create_schedule}
\alias{get_schedules}
\alias{get_schedule}
\alias{set_schedule_attributes}
\alias{delete_schedule}
\title{Hue API: schedules endpoints}
\usage{
create_schedule(..., return_id = FALSE)

get_schedules()

get_schedule(id)

set_schedule_attributes(id, ...)

delete_schedule(id)
}
\arguments{
\item{...}{named parameters describing schedule attributes
(e.g. \code{name = 'foo'})}

\item{return_id}{when creating a new resource, the default is to return
\code{TRUE} (invisibly) upon success; setting \code{return_id = TRUE} will
return the ID of the newly created resource instead.}

\item{id}{id of a specific schedule}
}
\value{
Requests with side effects return \code{TRUE} (invisibly) upon
  success. GET requests return the response content, parsed into a list.
}
\description{
Hue API: schedules endpoints
}
\seealso{
\url{https://developers.meethue.com/documentation/schedules-api-0}
}
