% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors_api.R
\name{sensors}
\alias{sensors}
\alias{create_sensor}
\alias{search_for_new_sensors}
\alias{get_new_sensors}
\alias{rename_sensor}
\alias{get_sensors}
\alias{get_sensor}
\alias{set_sensor_config}
\alias{set_sensor_state}
\alias{delete_sensor}
\title{Hue API: \code{sensors} endpoints}
\usage{
create_sensor(...)

search_for_new_sensors()

get_new_sensors()

rename_sensor(id, name)

get_sensors()

get_sensor(id)

set_sensor_config(id, ...)

set_sensor_state(id, ...)

delete_sensor(id)
}
\arguments{
\item{...}{named parameters describing sensor state (e.g. \code{on = TRUE})}

\item{id}{ID of a specific sensor}

\item{name}{name to assign to the sensor}
}
\value{
Requests that create resources return the ID of the newly created
item, requests with side effects return \code{TRUE} upon success, and GET
requests return the response content, parsed into a list.
}
\description{
Hue API: \code{sensors} endpoints
}
\seealso{
\url{https://developers.meethue.com/develop/hue-api/5-sensors-api/}
}
